GPS Deface Spy/backend/sentinel_grid/requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5
django-cors-headers>=4.3.0
apscheduler>=3.10.4
reportlab>=4.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
gunicorn>=21.2.0
numpy>=1.26.0
playwright>=1.40.0
